Use web preset to build webextension
* Renamed the preset to webextension
* Deleted unwanted configs from the web preset
* Store hot-update.{js, json} files in build/hot/
* Stop using hashes because it's not useful for
web extensions. Also, the main html and js files
need to be referenced in manifest.json and so it becomes
hard to reference them if they have hashes.
