--- packages/web-extension/package.json
@@ -1,40 +1,0 @@
-{
-  "name": "@neutrinojs/web-extension",
-  "version": "7.3.2",
-  "description": "Neutrino preset for building web extensions",
-  "main": "index.js",
-  "keywords": [
-    "neutrino",
-    "neutrino-preset",
-    "web-extension"
-  ],
-  "author": "Hassan Ali <helfi92@gmail.com>",
-  "license": "MPL-2.0",
-  "repository": "https://github.com/mozilla-neutrino/neutrino-dev/tree/master/packages/web-extension",
-  "homepage": "https://neutrino.js.org",
-  "bugs": "https://github.com/mozilla-neutrino/neutrino-dev/issues",
-  "dependencies": {
-    "@neutrinojs/chunk": "^7.3.2",
-    "@neutrinojs/clean": "^7.3.2",
-    "@neutrinojs/compile-loader": "^7.3.2",
-    "@neutrinojs/copy": "^7.3.2",
-    "@neutrinojs/env": "^7.3.2",
-    "@neutrinojs/font-loader": "^7.3.2",
-    "@neutrinojs/html-loader": "^7.3.2",
-    "@neutrinojs/html-template": "^7.3.2",
-    "@neutrinojs/image-loader": "^7.3.2",
-    "@neutrinojs/loader-merge": "^7.3.2",
-    "@neutrinojs/minify": "^7.3.2",
-    "@neutrinojs/style-loader": "^7.3.2",
-    "babel-plugin-syntax-dynamic-import": "^6.18.0",
-    "babel-preset-env": "^1.6.0",
-    "deepmerge": "^1.5.2",
-    "fast-async": "^6.3.0",
-    "webpack": "^3.6.0",
-    "webpack-manifest-plugin": "^1.3.2",
-    "web-ext": "^2.2.2"
-  },
-  "peerDependencies": {
-    "neutrino": "^8.0.0"
-  }
-}
